try to fix the coordinate of ground truth pose
With unknown reason, the coordinate of lidar pointcloud is under the normal xyz lidar coordinate.
However, the coordinate of ground truth pose is not under the same coordinate.
This can be validated by projecting lidar frames into world coordinate frames using the groudn truth pose.
The open source code seems to use a R_transform matrix to convert the ground truth pose into normal coordinate.
However it is still weird. This is because that the rotaiton matrix does not have 1 in the diagonal axis after the projection.